25. Processamento de dados em Java, Casting

+++ Aula25/src/Aula25.java
@@ -59,7 +59,7 @@ public class Aula25 {
 		int b;
 		
 		a = 5.0;
-		b = a;
+		b = (int)a;
 		
 		System.out.println(b);
 	}
